UPSTREAM: ath11k: Move parameters in bus_params to hw_params
In ath11k, bus_params were added with an intention to hold
parameters related to bus (AHB/PCI), but this is not true
as some bus parameters being different between chipsets of
the same bus.

With the addition of WCN6750 to ath11k, bus parameters are
going to be entirely different among AHB devices. Therefore,
it is wise to move bus_params to hw_params and get rid of
bus_params entirely.

Also, mhi_support parameter is not used anywhere in the driver,
remove it from bus_params.

Tested-on: WCN6750 hw1.0 AHB WLAN.MSL.1.0.1-00887-QCAMSLSWPLZ-1

Signed-off-by: Manikanta Pubbisetty <quic_mpubbise@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>
Link: https://lore.kernel.org/r/20220429170502.20080-3-quic_mpubbise@quicinc.com
(cherry picked from commit 92c1858e4399edc093a41cbf8f74874bdab59da7)

BUG=b:232096343
TEST=emerge-herobrine chromeos-kernel-5_15; deploy and test basic WiFi
     functionality on herobrine device

Signed-off-by: Kevin Lund <kglund@google.com>
Change-Id: I1ebc6adaa04c096685c1c46eb92c62bfa7427131
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3654380
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>