UPSTREAM: ath11k: Add HW params for WCN6750
WCN6750 is a PCIe based solution that is attached to and enumerated
by the WPSS (Wireless Processor SubSystem) Q6 processor.

Though it is a PCIe device, since it is not attached to APSS processor
(Application Processor SubSystem), APSS will be unaware of such a decice
and hence it is registered to the APSS processor as a platform device(AHB).
Because of this hybrid nature, it is called as a hybrid bus device.

A new variable hybrid_bus_type is defined in hw_params to indicate the
hybrid nature of the device.

Add HW params for WCN6750.

Tested-on: WCN6750 hw1.0 AHB WLAN.MSL.1.0.1-00887-QCAMSLSWPLZ-1
Tested-on: WCN6855 hw2.0 PCI WLAN.HSP.1.1-01720.1-QCAHSPSWPL_V1_V2_SILICONZ_LITE-1
Tested-on: QCN9074 hw1.0 PCI WLAN.HK.2.5.0.1-01100-QCAHKSWPL_SILICONZ-1
Tested-on: IPQ8074 hw2.0 AHB WLAN.HK.2.4.0.1-00192-QCAHKSWPL_SILICONZ-1

Signed-off-by: Manikanta Pubbisetty <quic_mpubbise@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>
Link: https://lore.kernel.org/r/20220429170502.20080-4-quic_mpubbise@quicinc.com
(cherry picked from commit d1e1edfde0352321af52599ad447b71c91bc6e76)

BUG=b:232096343
TEST=emerge-herobrine chromeos-kernel-5_15; deploy and test basic WiFi
     functionality on herobrine device

Signed-off-by: Kevin Lund <kglund@google.com>
Change-Id: Id47edb741cc66c0faf93aa18acfc02fb50b1ee54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3654381
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>