UPSTREAM: ath11k: Add QMI changes for WCN6750
In the case of WCN6750, FW doesn't request for DDR memory
via QMI, instead it uses a fixed 12MB reserved Memory region
in the DDR which is called as MSA region. As a result, QMI
message sequence is not same as other ath11k supported devices.

Also, M3 firmware will be bundled into the FW and will be
downloaded to the target as part of Q6 boot.

This is the QMI flow in the case of WCN6750,

1) QMI firmware indication REQ/RESP
2) QMI host capability REQ/RESP
3) QMI target capability REQ/RESP
4) QMI device info REQ/RESP
5) QMI BDF download
6) QMI FW ready

Tested-on: WCN6750 hw1.0 AHB WLAN.MSL.1.0.1-00887-QCAMSLSWPLZ-1
Tested-on: WCN6855 hw2.0 PCI WLAN.HSP.1.1-01720.1-QCAHSPSWPL_V1_V2_SILICONZ_LITE-1
Tested-on: QCN9074 hw1.0 PCI WLAN.HK.2.5.0.1-01100-QCAHKSWPL_SILICONZ-1
Tested-on: IPQ8074 hw2.0 AHB WLAN.HK.2.4.0.1-00192-QCAHKSWPL_SILICONZ-1

Signed-off-by: Manikanta Pubbisetty <quic_mpubbise@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>
Link: https://lore.kernel.org/r/20220429170502.20080-7-quic_mpubbise@quicinc.com
(cherry picked from commit 73d3e71306fe864d9667e8d37f731e93a91e2040)

BUG=b:232096343
TEST=emerge-herobrine chromeos-kernel-5_15; deploy and test basic WiFi
     functionality on herobrine device

Signed-off-by: Kevin Lund <kglund@google.com>
Change-Id: I73232dca92a05175a3c5187e3641eb2a1b97858a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3654384
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>