UPSTREAM: usb: typec: Separate USB Power Delivery from USB Type-C
Introducing a small device class for USB Power Delivery.
The idea with it is that we do not mix any more USB Power
Delivery information into the USB Type-C connectors only.
This separation will make it possible to register USB Power
Delivery devices also from other places, for example from
USB Type-C Bridges (see USB Type-C Bridge Specification).

The device class will not always deal with only the messages
and objects that were negotiated with the partner, but
instead messages and objects that can be used in the
negotiation. That allows the USB PD devices to be shared and
reconfigured. The ports can decide which objects are to be
advertised to the partner before the contract is negotiated.
It is also possible to allow the user space to make that
decision if needed.

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Link: https://lore.kernel.org/r/20220502132058.86236-2-heikki.krogerus@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 662a60102c122e44fdaf5c826f7f415eb57d48ad)

BUG=b:236280178
TEST=None

Signed-off-by: Prashant Malani <pmalani@chromium.org>
Change-Id: Ia1e73fff1fe5daf83a75fc2c568be4e3f513d157
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3811476
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Benson Leung <bleung@google.com>