FROMGIT: iommu/mediatek: Initialise/Remove for multi bank dev
The registers for each bank of the IOMMU base are in order, delta is
0x1000. Initialise the base for each bank.

For all the previous SoC, we only have bank0. thus use "do {} while()"
to allow bank0 always go.

When removing the device, Not always all the banks are initialised, it
depend on if there is masters for that bank.

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Reviewed-by: Matthias Brugger <matthias.bgg@gmail.com>
Link: https://lore.kernel.org/r/20220503071427.2285-34-yong.wu@mediatek.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit 42d57fc58aebc5801804424082028f43bad1b73c
 https://git.kernel.org/pub/scm/linux/kernel/git/joro/iommu.git arm/mediatek)

BUG=b:213021160
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15 succeeds

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Signed-off-by: Allen-kh Cheng <allen-kh.cheng@mediatek.corp-partner.google.com>
Change-Id: Idb98dfc00dda5b6332ea090bc9a1198665c3ab87
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3663851
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>