FROMGIT: ASoC: SOF: Add rx_data pointer to snd_sof_ipc_msg struct
The rx_data pointer can be used by IPC implementations to pass the received
message (or part of the message, like the header) from platform code to
generic, high level IPC code.

IPC4 is going to be the first user of this as its implementation on Intel
platforms detaches the header and payload and the rx cannot be handled in
a similar way as it is implemented for ipc3.

If the rx_data is dynamically allocated, it is up to the platform code to
free it up.
After the message reception handling (rx_msg ops) returned, the pointer via
the msg->rx_data should be considered as invalid.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Reviewed-by: Rander Wang <rander.wang@intel.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Link: https://lore.kernel.org/r/20220505094818.10346-2-peter.ujfalusi@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 84c2dca3c3d4b8e011a2536fc7aaf2a2bdc77972
 https://kernel.googlesource.com/pub/scm/linux/kernel/git/broonie/sound.git for-next)

BUG=b:218771967
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15

Signed-off-by: Tinghan Shen <tinghan.shen@mediatek.com>
Change-Id: I90c362c7003c0ed8a32aadd349760121d07a3981
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3653690
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>