FROMGIT: ASoC: SOF: ipc-msg-injector: Add support for IPC4 messages
The IPC message representation of an IPC4 differs from the IPC3 version
significantly.

The message for IPC4 should be written to the debugfs file in this form:
0-7 IPC4 header (2x u32)
8-  additional payload, if any

The reply is given back in the same form.

The message size limitation is the same as with the IPC3, only messages
which can fit to the mailbox can be injected (and received).

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Link: https://lore.kernel.org/r/20220506132647.18690-8-peter.ujfalusi@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 066c67624d8ca2a2465690d4a7b7f52b880e9925
 https://kernel.googlesource.com/pub/scm/linux/kernel/git/broonie/sound.git for-next)

BUG=b:218771967
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15

Signed-off-by: Tinghan Shen <tinghan.shen@mediatek.com>
Change-Id: I6b2aa9c474434f5bb61f7e3b0dfdb53ccd79e082
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3653699
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>