UPSTREAM: ath11k: Handle keepalive during WoWLAN suspend and resume
With WoWLAN enabled and after sleeping for a rather long time,
we are seeing that with some APs, it is not able to wake up
the STA though the correct wake up pattern has been configured.
This is because the host doesn't send keepalive command to
firmware, thus firmware will not send any packet to the AP and
after a specific time the AP kicks out the STA.

Fix this issue by enabling keepalive before going to suspend
and disabling it after resume back.

Tested-on: WCN6855 hw2.0 PCI WLAN.HSP.1.1-01720.1-QCAHSPSWPL_V1_V2_SILICONZ_LITE-1

Signed-off-by: Baochen Qiang <quic_bqiang@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>
Link: https://lore.kernel.org/r/20220506012540.1579604-1-quic_bqiang@quicinc.com
(cherry picked from commit 0f84a156aa3b9c9889c64a31d36b533508fabcb7)

BUG=b:248457431, b:204767474
TEST=emerge-herobrine chromeos-kernel-5_15; deploy and test basic WiFi
     functionalities on herobrine device.
TEST=suite:wifi_matfunc results are consistent with lab results.

Signed-off-by: Jun Yu <junyuu@chromium.org>
Change-Id: I8f763909ee03b9af4f34f81b07013adb4f10431b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4038507
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>