UPSTREAM: ath11k: Designating channel frequency when sending management frames
In case of Passpoint, the WLAN interface may be requested to
remain on a specific channel and then to send some management
frames on that channel. Now chanfreq of wmi_mgmt_send_cmd is set
as 0, as a result firmware may choose a default but wrong channel.
Fix it by assigning chanfreq field with the designated channel.

This change only applies to WCN6855 and QCA6390, other chips are
not affected.

Tested-on: WCN6855 hw2.0 PCI WLAN.HSP.1.1-01720.1-QCAHSPSWPL_V1_V2_SILICONZ_LITE-1

Signed-off-by: Baochen Qiang <quic_bqiang@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>
Link: https://lore.kernel.org/r/20220506013614.1580274-4-quic_bqiang@quicinc.com
(cherry picked from commit 1d7f514577f0ccf3e5f5736247138868fb62896a)

BUG=b:248457431, b:177043467
TEST=emerge-herobrine chromeos-kernel-5_15; deploy and test basic WiFi
     functionalities on herobrine device.

Signed-off-by: Jun Yu <junyuu@chromium.org>
Change-Id: Id9f3f5c62ac4a0d4bf41d84bc88112883fb09084
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3943174
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>