UPSTREAM: serial: core: Do stop_rx in suspend path for console if console_suspend is disabled
For the case of console_suspend disabled, if back to back suspend/resume
test is executed, at the end of test, sometimes console would appear to
be frozen not responding to input. This would happen because, during
resume, rx transactions can come in before system is ready, malfunction
of rx happens in turn resulting in console appearing to be stuck.

Do a stop_rx in suspend sequence to prevent this.

Signed-off-by: Vijaya Krishna Nivarthi <quic_vnivarth@quicinc.com>
Link: https://lore.kernel.org/r/1652692810-31148-1-git-send-email-quic_vnivarth@quicinc.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit c9d2325cdb92fd4a6362ea792d93571195741675)

BUG=b:211843408
TEST=suspend/resume

Cq-Depend: chromium:3699363
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ifc8da11454b223fa22aabf1aec1d58b7a4e39a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3699361
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>