FROMGIT: ASoC: SOF: ipc-msg-injector: Cap the rmaining to count in IPC4 mode
If user space provides smaller buffer than the IPC4 reply then it is
possible that we corrupt user space memory since the IPC4 dfs_read function
is not using the count directly in copy_to_user() due to the nature of
an IPC4 message.

Cap the remaining counter to make sure that we are not writing too much to
the user space provided buffer.

Add a check also to make sure that the buffer is at least the size of the
IPC4 header.

Fixes: 066c67624d8c: "ASoC: SOF: ipc-msg-injector: Add support for IPC4 messages"
Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Link: https://lore.kernel.org/r/20220516092442.17027-1-peter.ujfalusi@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit c5003f08fe671fb1f18bca07e589c5cffeccbc9b
 https://kernel.googlesource.com/pub/scm/linux/kernel/git/broonie/sound.git for-next)

BUG=b:218771967
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15

Signed-off-by: Tinghan Shen <tinghan.shen@mediatek.com>
Change-Id: Ibcc6488822196217983db07e8b3d33621e972c99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3653712
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>