CHROMIUM: iwl7000: mvm: refactor iwl_mvm_add_sta(), iwl_mvm_rm_sta()
As we have a new MLD STA cmd, there will be a different function to
add/remove
a station in MLD mode. But both functions will share a common part.
Put this part in a separate function which will be used later in
adding/removing a sta in MLD mode.

Change-Id: I5e9e27c3b363b36209a0ff960d2e59708e7ff0bf
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 7978c201094864a5e9d05e2e5ac85cbdf3335731