FROMGIT: ASoC: core: Always send the CPU DAI a direct clock specifier
All CPU drivers are now updated to accept a direct indication of
whether they are clock provider or consumer, rather than being told
if the CODEC is clock provider. As such update the core to always
pass this direct indication rather than only if the new set_fmt_new
callback is defined. Note this makes no difference to the CODEC
side of the DAI link as it was already being directly told if it
was clock provider under the old system.

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20220519154318.2153729-29-ckeepax@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 6c076273a326cc5b5162451aacf7b7744bb03c66
 http://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-next)

BUG=b:218771967
TEST=boot to shell

Signed-off-by: Chunxu Li <chunxu.li@mediatek.corp-partner.google.com>
Change-Id: I787c6bb3d44c0ae42618e13bf50f9870e1243a2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3721785
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>