UPSTREAM: net: ipa: kill gsi_trans_commit_wait_timeout()
Since the beginning gsi_trans_commit_wait_timeout() has existed to
provide a way to allow waiting a limited time for a transaction
to complete.  But that function has never been used.

In fact, there is no use for this function, because a transaction
committed to hardware should *always* complete.  The only reason it
might not complete is if there were a hardware failure, or perhaps a
system configuration error.

Furthermore, if a timeout ever did occur, the IPA hardware would be
in an indeterminate state, from which there is no recovery.  It
would require some sort of complete IPA reset, and would require the
participation of the modem, and at this time there is no such
sequence defined.

So get rid of the definition of gsi_trans_commit_wait_timeout(), and
update a few comments accordingly.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit d15180b4eadbdee782d97afcbf3c56e2f246c40a)

BUG=b:228884730
TEST=Tested on Lazor. Wifi, LTE, and bluetooth all still work.

Signed-off-by: Joseph S. Barrera III <joebar@chromium.org>
Change-Id: Iaf87c08a7e9c253c7f9ae7cea67778349b112849
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3774983
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>