UPSTREAM: net: ipa: get rid of ipa_cmd_info->direction
The direction field of the ipa_cmd_info structure is set, but never
used.  It seems it might have been used for the DMA_SHARED_MEM
immediate command, but the DIRECTION flag is set based on the value
of the passed-in direction flag there.

Anyway, remove this unused field from the ipa_cmd_info structure.
This is done as a separate patch to make it very obvious that it's
not required.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 7ffba3bdf76a7bf93ec2dc56369ea563dea92b7c)

BUG=b:228884730
TEST=Tested on Lazor. Wifi, LTE, and bluetooth all still work.

Signed-off-by: Joseph S. Barrera III <joebar@chromium.org>
Change-Id: Iac8313ad1ddb299fa0c3da37a55c8d6ca923ad57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3774985
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>