UPSTREAM: net: ipa: remove command info pool
The ipa_cmd_info structure now contains only one field, and it's an
enumerated type whose values all fit in 8 bits.  Currently we'll
never use more than 8 TREs in a command transaction, and we can
represent that number of command opcodes in the same space as a 64
bit pointer to an ipa_cmd_info structure.

Define IPA_COMMAND_TRANS_TRE_MAX as the maximum number of TREs that
can be in a command transaction.  Replace the info pointer in a
transaction with a fixed-size array named cmd_opcode[] of that many
bytes.  Store the opcode in this array when adding a command TRE to
a transaction, as was done previously for the info array.  This
makes the ipa_cmd_info unused, so get rid of it.

When committing an immediate command transaction, use the channel's
Boolean command flag to determine whether to fill in the opcode,
which will be taken (as before) from the array in the transaction.

This makes the command info pool unnecessary, so get rid of it.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 8797972afff3d462711758acec68be11172e2d01)

BUG=b:228884730
TEST=Tested on Lazor. Wifi, LTE, and bluetooth all still work.

Signed-off-by: Joseph S. Barrera III <joebar@chromium.org>
Change-Id: I0b1cdaa0b9d28378f929b9f5be9a86b6a1e52cc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3774987
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>