UPSTREAM: drm/mediatek: Fix DPI component detection for MT8192
When support for MT8192 was added, the DPI device was not added to the
list of components to look for. This causes the secondary display
pipeline to not be able to fully bind, and the DRM driver subsequently
defers probing.

Add the DPI device compatible to list of DPI components to fix this.

Link: https://patchwork.kernel.org/project/linux-mediatek/patch/20220225032754.140168-1-wenst@chromium.org/
Fixes: 01365f549c88 ("drm/mediatek: Add support for Mediatek SoC MT8192")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Signed-off-by: Chun-Kuang Hu <chunkuang.hu@kernel.org>
(cherry picked from commit cfab37ff31afcd0f99f3cccbff1f8ffa11e44c00)

BUG=b:204164699
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: I5ed4893f316a791d28991fc7c33fd1e16929cb18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3670538
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: Hsin-Yi Wang <hsinyi@chromium.org>