UPSTREAM: drm/vc4: kms: Take old state core clock rate into account
During a commit, the core clock, which feeds the HVS, needs to run at
a minimum of 500MHz.

While doing that commit, we can also change the mode to one that
requires a higher core clock, so we take the core clock rate associated
to that new state into account for that boost.

However, the old state also needs to be taken into account if it
requires a core clock higher that the new one and our 500MHz limit,
since it's still live in hardware at the beginning of our commit.

Fixes: 16e101051f32 ("drm/vc4: Increase the core clock based on HVS load")
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Acked-by: Thomas Zimmermann <tzimmermann@suse.de>
Link: https://lore.kernel.org/r/20220331143744.777652-2-maxime@cerno.tech
(cherry picked from commit 748acfc98adab21a93ae7a1b5bed0f048463e873)

BUG=b:227338908
TEST=build test

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: I6988dc9a9a4253e4c44b50097c1bf34ab49ccbf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3670539
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>