UPSTREAM: thunderbolt: Add CL1 support for USB4 and Titan Ridge routers
In this patch we add support for a second low power state of the link: CL1.
Low power states (called collectively CLx) are used to reduce
transmitter and receiver power when a high-speed lane is idle.
We enable it, if both sides of the link support it,
and only for the first hop router (i.e. the first device that connected
to the host router). This is needed for better thermal management.

Signed-off-by: Gil Fine <gil.fine@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit b017a46d486cd4113b1856f3fd611f54cd0f9c03)

BUG=b:242112030
TEST=Verify TBT3/USB4 enumeration with USB/DP/PCIe tunnels working.

Change-Id: I24b3a24a97de58b29d0917050b1176c9ce0f2c7e
Signed-off-by: Utkarsh Patel <utkarsh.h.patel@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4063222
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>